Cleaned up code and fixed next build
